Moves all server logic from MysportsfeedController to UpdateDB job.
